Additional fixes for es reorganization:
Fix async d.ts files being broken
Fix es6 module output not transpiling async generators
Give split index into asnyc index and non-async index
Fix package.json main, module, and typings
Run build on prepack: will make some aspects of the library work when included as git ref
