+++ package.json
@@ -11,8 +11,10 @@
     "watch": "npm run test -- -w",
     "lint": "eslint --fix --ext .mjs ./src ./tests ./benchmarks",
     "clean": "rm -rf lib/*.* && rm -rf lib/async/*.* && rm -rf es5/*.* && rm -rf es5/async/*.* && rm -rf package/*",
-    "build": "npm run clean && cross-env BABEL_ENV=es5 babel -x \".mjs\" src -d es5 && cross-env BABEL_ENV=cjs babel -x \".mjs\" src -d lib && cp src/*.mjs src/*.d.ts lib && cp src/async/*.mjs src/async/*.d.ts lib/async",
-    "dist": "npm run build && cp -R package.json lib/* es5 LICENSE README.md package/ && cd package && npm publish",
+    "build": "npm run clean && cross-env BABEL_ENV=es6 babel -x \".mjs\" src -d lib && renamer --regex --find '\\.js$' --replace '.mjs' 'lib/**' && cross-env BABEL_ENV=es5 babel -x \".mjs\" src -d es5 && cross-env BABEL_ENV=cjs babel -x \".mjs\" src -d lib && cp src/*.d.ts lib && cp src/async/*.mjs src/async/*.d.ts lib/async",
+    "dist": "npm run build && cp -R package.json lib/* es5 LICENSE README.md package/ && cd package && json -I -f package.json -e \"this.main='index.js';this.typings='index.d.ts';this.module='index.mjs'\" && npm publish",
+    "prepare": "#https://github.com/npm/npm/issues/19564",
+    "prepack": "npm run build",
     "precommit": "npm run lint",
     "prepush": "npm run test"
   },
@@ -45,9 +47,10 @@
     "eslint-plugin-promise": "^3.8.0",
     "eslint-plugin-standard": "^3.1.0",
     "husky": "^0.14.3",
+    "json": "^9.0.6",
     "measure-speed": "^1.0.0",
     "mocha": "^5.0.5",
-    "renamer": "^0.6.1",
+    "renamer": "0.5.7",
     "require-all": "^2.2.0"
   },
   "engines": {
