+++ src/async/compress.d.ts
@@ -1,2 +1,2 @@
-import { asyncCompress } from '../../index';
+import { asyncCompress } from './index';
 export = asyncCompress;
