+++ src/async/tee.d.ts
@@ -1,2 +1,2 @@
-import { asyncTee } from '../../index';
+import { asyncTee } from './index';
 export = asyncTee;
