+++ src/index.mjs
@@ -28,24 +28,3 @@ export { default as takeWhile } from './take-while'
 export { default as tee } from './tee'
 export { default as zipLongest } from './zip-longest'
 export { default as zip } from './zip'
-export { default as asyncIterToArray } from './async/async-iter-to-array'
-export { default as asyncIter } from './async/async-iter'
-export { default as asyncChain } from './async/chain'
-export { default as asyncCompress } from './async/compress'
-export { default as asyncCycle } from './async/cycle'
-export { default as asyncDropWhile } from './async/drop-while'
-export { default as asyncEnumerate } from './async/enumerate'
-export { default as asyncExecute } from './async/execute'
-export { default as asyncFilter } from './async/filter'
-export { default as asyncFlatMap } from './async/flat-map'
-export { default as asyncGroupBy } from './async/group-by'
-export { default as asyncMap } from './async/map'
-export { default as asyncReduceIter } from './async/reduce-iter'
-export { default as asyncReduce } from './async/reduce'
-export { default as asyncSlice } from './async/slice'
-export { default as asyncTakeWhile } from './async/take-while'
-export { default as asyncTee } from './async/tee'
-export { default as asyncZipLongest } from './async/zip-longest'
-export { default as asyncZip } from './async/zip'
-export { default as asyncRegexpSplitIter } from './async/regexp-split-iter'
-export { default as asyncSplitLines } from './async/split-lines'
